ArtBook Project
In this project, a travel book was created for the user. Since the user needed to use an image from their gallery, permission to access photos was requested. After the required fields were filled in by the user, when the “SAVE” button was pressed, the necessary historical artifact name, artist, and year were saved into a RecyclerView within a separate Activity. Additionally, the data in the list was stored in a local database, ensuring that previous records remained accessible even if the application was closed and reopened. Multi-screen support was also provided for the project. I would like to express my gratitude to my esteemed instructor, Atil SAMANCIOGLU, for the valuable knowledge I have gained.