ArtBook with Navigations & Room
This project is exactly the same as the “ArtBook” project I previously developed. However, the technologies used in it are different. In the earlier version, I built the project within an Activity using plain SQLite. In this version, I made the project more modern and up-to-date by using Navigation components and integrating the local SQLite database with the Room library.